catalog: add GetKubecontroller interface
XDS verticals have no way to call some of the existing functions
(namely, GetPodFromCertificate).

To avoid creating an API everytime we have a use-case that needs
special massaging of data, this commit allows packages to get
visibility on the kubecontroller instance to allow verticals see
kubernetes resources directly if they need to.

Signed-off-by: Eduard Serra <eduser25@gmail.com>